% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_attach_to_uuid}
\alias{check_arg_attach_to_uuid}
\title{Check argument attach_to_uuid}
\usage{
check_arg_attach_to_uuid(attach_to_uuid)
}
\arguments{
\item{attach_to_uuid}{argument to check}
}
\value{
TRUE or error
}
\description{
Attach to uuid has not been implemented yet and hence if not null is now considered FALSE
}
