% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_char_opts}
\alias{check_arg_char_opts}
\title{Check argument in character options}
\usage{
check_arg_char_opts(object, opts, null = FALSE)
}
\arguments{
\item{object}{argument to check}

\item{opts}{Possible values for object}

\item{null}{Whether NULL value is allowed, default is FALSE}
}
\value{
TRUE or error
}
\description{
Check if argument is in the given options
}
