% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_dataframe}
\alias{check_arg_dataframe}
\title{Check argument data frame}
\usage{
check_arg_dataframe(dataframe)
}
\arguments{
\item{dataframe}{argument to check}
}
\value{
TRUE or error
}
\description{
Dataset should be a data frame
}
