% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_dataset_id}
\alias{check_arg_dataset_id}
\title{Check argument dataset_id}
\usage{
check_arg_dataset_id(dataset_id)
}
\arguments{
\item{dataset_id}{argument to check}
}
\value{
TRUE or error
}
\description{
Dataset id should be NULL or a numeric and indicate (an) existing dataset(s)
}
