% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_description}
\alias{check_arg_description}
\title{Check argument description}
\usage{
check_arg_description(description)
}
\arguments{
\item{description}{argument to check}
}
\value{
TRUE or error
}
\description{
Description should be NULL or a single character and is used as a description for an object on Labguru
}
