% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_email}
\alias{check_arg_email}
\title{Check argument email}
\usage{
check_arg_email(email)
}
\arguments{
\item{email}{argument to check}
}
\value{
TRUE or error
}
\description{
Email should be a single character and represent a valid email address
}
