% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_file}
\alias{check_arg_file}
\title{Check argument file}
\usage{
check_arg_file(file)
}
\arguments{
\item{file}{argument to check}
}
\value{
TRUE or error
}
\description{
File should be a single character and represent a file that actually exists.
}
