% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_get_cols}
\alias{check_arg_get_cols}
\title{Check argument get_cols}
\usage{
check_arg_get_cols(get_cols, opts)
}
\arguments{
\item{get_cols}{argument to check}

\item{opts}{Possible values for get_cols}
}
\value{
TRUE or error
}
\description{
Character with columns of a dataframe to return
}
