% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_name}
\alias{check_arg_name}
\title{Check argument name}
\usage{
check_arg_name(name)
}
\arguments{
\item{name}{argument to check}
}
\value{
TRUE or error
}
\description{
File should be a single character and is used as name for a visualisation on Labguru
}
