% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_numeric}
\alias{check_arg_numeric}
\title{Check argument numeric}
\usage{
check_arg_numeric(object, null = FALSE)
}
\arguments{
\item{object}{object to check}

\item{null}{Whether NULL value is allowed, default is FALSE}
}
\value{
TRUE or error
}
\description{
Object should be a numeric
}
