% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_page}
\alias{check_arg_page}
\title{Check argument page}
\usage{
check_arg_page(page)
}
\arguments{
\item{page}{argument to check}
}
\value{
TRUE or error
}
\description{
Page number of GET request must be integer
}
