% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_password}
\alias{check_arg_password}
\title{Check argument password}
\usage{
check_arg_password(password)
}
\arguments{
\item{password}{argument to check}
}
\value{
TRUE or error
}
\description{
Password should be a single character
}
