% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_server}
\alias{check_arg_server}
\title{Check argument server}
\usage{
check_arg_server(server)
}
\arguments{
\item{server}{argument to check}
}
\value{
TRUE or error
}
\description{
Server is the base URL for the API connection and has to be a character string of length 1.
}
