% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_set_sys}
\alias{check_arg_set_sys}
\title{Check argument set_sys}
\usage{
check_arg_set_sys(set_sys)
}
\arguments{
\item{set_sys}{argument to check}
}
\value{
TRUE or error
}
\description{
Set sys should be a single logical
}
