% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_single_logical}
\alias{check_arg_single_logical}
\title{Check argument single logical}
\usage{
check_arg_single_logical(object, null = FALSE)
}
\arguments{
\item{object}{object to check}

\item{null}{Whether NULL value is allowed, default is FALSE}
}
\value{
TRUE or error
}
\description{
Password should be a single logical
}
