% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_title}
\alias{check_arg_title}
\title{Check argument title}
\usage{
check_arg_title(title)
}
\arguments{
\item{title}{argument to check}
}
\value{
TRUE or error
}
\description{
File should be a single character and is used as title for an object on Labguru
}
