% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_arg_token}
\alias{check_arg_token}
\title{Check argument token}
\usage{
check_arg_token(token)
}
\arguments{
\item{token}{argument to check}
}
\value{
TRUE or error
}
\description{
Token is the base URL for the API connection and has to be a character string of length 1
}
