% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{ic50_create_control_data}
\alias{ic50_create_control_data}
\title{Labguru create control data for ic50}
\usage{
ic50_create_control_data(df, wells)
}
\arguments{
\item{df}{extracted plate data}

\item{wells}{extracted wells data}
}
\value{
data frame of control data
}
\description{
Labguru create control data for ic50
}
