% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{ic50_create_dilution_data}
\alias{ic50_create_dilution_data}
\title{Labguru create dilution data for ic50}
\usage{
ic50_create_dilution_data(df)
}
\arguments{
\item{df}{extracted plate data}
}
\value{
data frame of dilution data
}
\description{
Labguru create dilution data for ic50
}
