% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{ic50_create_measure_data}
\alias{ic50_create_measure_data}
\title{Labguru create measure data for ic50}
\usage{
ic50_create_measure_data(df)
}
\arguments{
\item{df}{extracted plate data}
}
\value{
data frame of measure data
}
\description{
Labguru create measure data for ic50
}
