% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{ic50_extract_data_from_json}
\alias{ic50_extract_data_from_json}
\title{ic50 extract data from json}
\usage{
ic50_extract_data_from_json(prs_data)
}
\arguments{
\item{data}{Parsed JSON plate data from Labguru}
}
\value{
data frame
}
\description{
ic50 extract data from json
}
