% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_valid_email}
\alias{is_valid_email}
\title{Is valid email}
\usage{
is_valid_email(x)
}
\arguments{
\item{x}{email address

Check if an email address is valid}
}
\value{
BOOLEAN
}
\description{
Is valid email
}
\examples{
is_valid_email("my@valid_email.com")
is_valid_email("myinvalid.email@com")
}
