% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{labguru_add_element}
\alias{labguru_add_element}
\title{Labguru add element}
\usage{
labguru_add_element(
  data = NULL,
  img_id = NULL,
  rscript = NULL,
  experiment_procedure_id,
  return = "id",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{data}{character(1) custom html to add or NULL (default),}

\item{img_id}{numeric(1) indicating the image (attachment) id on the labguru server or NULL (default),}

\item{rscript}{character(1) file name of an .R file (or .txt file) or NULL (default),}

\item{experiment_procedure_id}{numeric(1) The experiment procedure id for which to add a new element}

\item{return}{character(1) whether the function returns either 'id' (default) or 'all' element information}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list with either element id only or all element information
}
\description{
Add a new element to labguru
}
\examples{
\dontrun{
labguru_add_element(data                    = "<p>Hello World</p>",
                    experiment_procedure_id = 1,
                    return                  = "id")
}
}
