% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{labguru_add_experiment}
\alias{labguru_add_experiment}
\title{Labguru add experiment}
\usage{
labguru_add_experiment(
  title,
  project_id,
  folder_id,
  description = NULL,
  return = "id",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{title}{character(1) The title of the experiment}

\item{project_id}{numeric(1) The project id for which to add a new experiment}

\item{folder_id}{numeric(1) The folder id for which to add a new experiment}

\item{description}{character(1) The description of the experiment}

\item{return}{character(1) whether the function returns either 'id' (default) or 'all' experiment information}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list with either experiment id only or all experiment information
}
\description{
Add a new experiment to labguru
}
\examples{
\dontrun{
labguru_add_experiment(project_id  = 1,
                       title       = "My new experiment",
                       description = "This experiment contains ...")
}
}
