% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_procedures.R
\name{labguru_add_experiment_procedure}
\alias{labguru_add_experiment_procedure}
\title{Labguru add experiment procedure}
\usage{
labguru_add_experiment_procedure(name, experiment_id, return = "id",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{name}{character(1) The name of the experiment procedure}

\item{experiment_id}{numeric(1) The experiment id for which to add a new procedure}

\item{return}{character(1) whether the function returns either 'id' (default) or 'all' experiment procedure information}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list with either experiment procedure id only or all experiment procedure information
}
\description{
Add a new experiment procedure to labguru
}
\examples{
\dontrun{
labguru_add_experiment_procedure(name          = "My new experiment procedure",
                                 experiment_id = 1,
                                 return        = "id")
}
}
