% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{labguru_add_project}
\alias{labguru_add_project}
\title{Labguru add project}
\usage{
labguru_add_project(
  title,
  description = NULL,
  return = "id",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{title}{character(1) The title of the project}

\item{description}{character(1) The description of the project}

\item{return}{character(1) whether the function returns either 'id' (default) or 'all' project information}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list with either project id only or all project information
}
\description{
Add a new project to labguru
}
\examples{
\dontrun{
labguru_upload_project(title       = "My new project",
                       description = "This project is an analysis of ...")
}
}
