% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{labguru_authenticate}
\alias{labguru_authenticate}
\title{Labguru Authenticate}
\usage{
labguru_authenticate(
  email,
  password,
  server = "https://my.labguru.com",
  set_sys = TRUE
)
}
\arguments{
\item{email}{character(1) email address}

\item{password}{character(1) password}

\item{server}{character(1) Server URL, "https://jonathan.labguru.com" by default}

\item{set_sys}{logical(1) Set server and token as system variables LABGURU_SERVER and LABGURU_TOKEN}
}
\value{
acces token (string)
}
\description{
This function authenticates a user by email and password and returns a token and sets the system
variable LABGURU_TOKEN to that token by default.
}
\examples{
\dontrun{
labguru_authenticate(email    = "my@email.com",
                     password = "mypassword",
                     server   = "https://my.labguru.com")
}
}
