% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{labguru_download_plate}
\alias{labguru_download_plate}
\title{Labguru download plate}
\usage{
labguru_download_plate(
  plate,
  dir = "./plate",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{plate}{numeric(1) indicating the plate id}

\item{dir}{character(1) indicating the directory to store the created files}

\item{server}{character(1) string indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
invisible, list() with dir name, and dilution, measure nd control data frames
}
\description{
Download a plate's data and stores dilution, measure and control files in a folder or returns data frames.
}
\examples{
\dontrun{
labguru_download_plate(plate = 271)
}
}
