% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api_functions.R
\name{labguru_get_by_id}
\alias{labguru_get_by_id}
\title{Labguru get by id}
\usage{
labguru_get_by_id(
  type,
  id,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{type}{character(1) type name of what you want to request}

\item{id}{numeric(1) id indicating a experiment on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
parsed object
}
\description{
Takes a type name and id and gets the information from the API.
}
\examples{
\dontrun{
labguru_get_by_id(type = "experiments",
                  id   = 1)
}
}
