% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{labguru_get_element}
\alias{labguru_get_element}
\title{Labguru get element}
\usage{
labguru_get_element(
  element_id,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{element_id}{numeric(1) id indicating a element on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list object of labguru element
}
\description{
Takes a element id and gets the element information.
}
\examples{
\dontrun{
labguru_get_element(element_id = 1)
}
}
