% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{labguru_get_experiment}
\alias{labguru_get_experiment}
\title{Labguru get experiment}
\usage{
labguru_get_experiment(experiment_id, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{experiment_id}{numeric(1) id indicating a experiment on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list object of labguru experiment
}
\description{
Takes a experiment id and gets the experiment information.
}
\examples{
\dontrun{
labguru_get_experiment(experiment_id = 1)
}
}
