% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_procedures.R
\name{labguru_get_experiment_procedure}
\alias{labguru_get_experiment_procedure}
\title{Labguru get experiment procedure}
\usage{
labguru_get_experiment_procedure(
  experiment_procedure_id,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{experiment_procedure_id}{numeric(1) id indicating a experiment procedure on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list object of labguru experiment procedure
}
\description{
Takes a experiment procedure id and gets the experiment procedure information.
}
\examples{
\dontrun{
labguru_get_experiment_procedure(experiment_procedure_id = 1)
}
}
