% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{labguru_get_folder}
\alias{labguru_get_folder}
\title{Labguru get folder}
\usage{
labguru_get_folder(folder_id, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{folder_id}{numeric(1) id indicating a folder on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list object of labguru folder
}
\description{
Takes a folder id and gets the folder information.
}
\examples{
\dontrun{
labguru_get_folder(folder_id = 1)
}
}
