% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{labguru_get_project}
\alias{labguru_get_project}
\title{Labguru get project}
\usage{
labguru_get_project(
  project_id,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{project_id}{numeric(1) id indicating a project on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
list of labguru project
}
\description{
Takes a project id and gets the project with milestones.
}
\examples{
\dontrun{
labguru_get_project(project_id = 1)
}
}
