% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{labguru_ic50_analysis}
\alias{labguru_ic50_analysis}
\title{ic50 Analysis}
\usage{
labguru_ic50_analysis(plate_id, indir = ".", outdir_plate = "./plate",
  outdir_results = "./results", plates = 2, inhib = NULL,
  normalize = "single", graphics = "mean", img_png = TRUE,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{plate_id}{}

\item{indir}{}

\item{outdir_plate}{}

\item{outdir_results}{}

\item{plates}{}

\item{inhib}{}

\item{normalize}{}

\item{graphics}{}

\item{img_png}{}

\item{server}{}

\item{token}{}
}
\description{
Download plate from labguru server and run ic50 analysis.
}
