% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_link_visualization}
\alias{labguru_link_visualization}
\title{Labguru link visualization}
\usage{
labguru_link_visualization(
  dataset_id,
  attachment_id,
  name,
  description = NULL,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{dataset_id}{numeric(1) The dataset id for which to link visualization}

\item{attachment_id}{numeric(1) The attachment id for which to link visualization}

\item{name}{character(1) the name}

\item{description}{character(1) The description of the folder}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{

}
\description{
Link a visualization to a dataset on the Labguru server.
}
\examples{
# helper for visulization linkage
}
