% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{labguru_list_datasets}
\alias{labguru_list_datasets}
\title{Labguru list datasets}
\usage{
labguru_list_datasets(page = 1, get_cols = "limited",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{page}{numeric(1) representing the page number of data to request. Limited data can be return in 1 request, incrementally try higher page numbers for more datasets}

\item{get_cols}{character(1) either 'limited' or 'all' to return a subset or all of the information regarding the datasets}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
dataframe with information of datasets, NULL if no datasets were available for the request
}
\description{
This function returns information of the available datasets in a data frame.
}
\examples{
\dontrun{
labguru_list_datasets()
}
}
