% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{labguru_list_elements}
\alias{labguru_list_elements}
\title{Labguru list elements}
\usage{
labguru_list_elements(experiment_id = NULL, page = 1,
  get_cols = "limited", server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{experiment_id}{numeric(1) The experiment id for which to list elements.}

\item{page}{numeric(1) representing the page number of data to request. Limited data can be return in 1 request, incrementally try higher page numbers for more elements}

\item{get_cols}{character(1) either 'limited' or 'all' to return a subset or all of the information regarding the elements}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
dataframe with information of elements, NULL if no elements were available for the request
}
\description{
This function returns information of the available elements in a data frame.
}
\examples{
\dontrun{
# The following example shows limited information for elements in all elements (default)
labguru_list_elements(experiment_id = NULL, 
                      page          = 1, 
                      get_cols      = "limited")
}
}
