% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{labguru_list_experiments}
\alias{labguru_list_experiments}
\title{Labguru list experiments}
\usage{
labguru_list_experiments(folder_id = NULL, page = 1, get_cols = "limited",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{folder_id}{numeric(1) The folder id for which to list experiments.}

\item{page}{numeric(1) representing the page number of data to request. Limited data can be return in 1 request, incrementally try higher page numbers for more experiments}

\item{get_cols}{character(1) either 'limited' or 'all' to return a subset or all of the information regarding the experiments}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
dataframe with information of experiments, NULL if no projects were available for the request
}
\description{
This function returns information of the available experiments in a data frame.
}
\examples{
\dontrun{
# The following example shows limited information for experiments in all folders (default)
labguru_list_experiments(folder_id = NULL, 
                         page      = 1, 
                         get_cols  = "limited")
}
}
