% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{labguru_list_folders}
\alias{labguru_list_folders}
\title{Labguru list folders}
\usage{
labguru_list_folders(
  project_id = NULL,
  page = 1,
  get_cols = "limited",
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{project_id}{numeric(1) The project is for which to list folders, NULL (default) returns for all projects}

\item{page}{numeric(1) representing the page number of data to request. Limited data can be return in 1 request, incrementally try higher page numbers for more folders}

\item{get_cols}{character(1) either 'limited' or 'all' to return a subset or all of the information regarding the folders}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
dataframe with information of folders, NULL if no projects were available for the request
}
\description{
This function returns information of the available projects in a data frame.
}
\examples{
\dontrun{
labguru_list_folders(project_id = NULL, page = 1, get_cols = "limited") # shows limited information for folders in all columns (default)
}
}
