% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api_functions.R
\name{labguru_list_items}
\alias{labguru_list_items}
\title{Labguru list items}
\usage{
labguru_list_items(url, config = NULL, handle = NULL)
}
\arguments{
\item{url}{url to the labGuru instance API endpoint}

\item{config}{other configuration options passed on to httr GET}

\item{handle}{handling arguments passed on to httr POST}
}
\value{
parsed object
}
\description{
List
}
\examples{
# helper function for GET requests
}
