% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api_functions.R
\name{labguru_post_item}
\alias{labguru_post_item}
\title{Labguru post item}
\usage{
labguru_post_item(url, body, encode = NULL, config = NULL, handle = NULL)
}
\arguments{
\item{url}{}

\item{body}{}

\item{encode}{}

\item{config}{}

\item{handle}{}
}
\value{
parsed object
}
\description{
POST a body to a url with optional arguments. Function includes standard error handling.
}
