% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elements.R
\name{labguru_read_rscript_to_html}
\alias{labguru_read_rscript_to_html}
\title{Labguru read rscript to html}
\usage{
labguru_read_rscript_to_html(file)
}
\arguments{
\item{file}{Path to a .R file, or .txt file}
}
\value{
character(1) paragraphed R script
}
\description{
Read an R script and place each line into paragraph tags (<p>)
}
\examples{
# helper function to parse to html
}
