% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{labguru_set_token}
\alias{labguru_set_token}
\title{Labguru set token}
\usage{
labguru_set_token(token, server = "https://jonathan.labguru.com",
  set_sys = TRUE)
}
\arguments{
\item{token}{Labguru API access token}

\item{server}{character(1) Server URL, "https://jonathan.labguru.com" by default}

\item{set_sys}{logical(1) Set server and token as system variables LABGURU_SERVER and LABGURU_TOKEN}
}
\value{
Returns the token (visibly or invisibly depending on set_sys)
}
\description{
Set token and server as system variables. Useful when a user has a token but no login credentials.
}
