% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{labguru_sys_variables}
\alias{labguru_sys_variables}
\title{Labguru system variables}
\usage{
labguru_sys_variables()
}
\value{
print Labguru system variables
}
\description{
Labguru system variables
}
\examples{
\dontrun{
labguru_sys_variables()
}
}
