% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{labguru_upload_dataset}
\alias{labguru_upload_dataset}
\title{Labguru Upload Dataset}
\usage{
labguru_upload_dataset(dataset, name, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{dataset}{An R data frame}

\item{name}{A character string}

\item{server}{A character string indicating the server URL}

\item{token}{An access token for API authentication}

\item{description}{A character string or NULL (default) to ignore}
}
\description{
Upload a dataset to labguru
}
\examples{
\dontrun{
labguru_upload_dataset(dataset    = mtcars,
                       name       = "mtcars",
                       desription = "Motor Trend Car Road Tests data")
}
}
