% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_upload_file}
\alias{labguru_upload_file}
\title{Labguru upload file}
\usage{
labguru_upload_file(file_path, name, description = NULL,
  attach_to_uuid = NULL, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{file_path}{String - File path of the file}

\item{name}{String - Name of the file to use on Labguru}

\item{description}{(optional) String - Text description for this attachment}

\item{attach_to_uuid}{(optional) String - The LG_UUID of the object to which this file should be attached}

\item{server}{A character string indicating the server URL}

\item{token}{An access token for API authentication}
}
\description{
Upload a file to Labguru
}
\examples{
data("compensation")

library(ggplot2)
plot <- ggplot(compensation,aes(x = Root, y = Fruit, colour = Grazing)) +
  geom_point(size = 5) +
  xlab("Root Biomass") +
  ylab("Fruit Production") +
  theme_bw()
  
ggsave("grazing.png", plot)

labguru_upload_image(file_path   = "grazing.png",
                     name        = "Grazing",
                     description = "Fruit production versus Root biomass for grazed and ungrazed")
}
