% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_upload_visualization}
\alias{labguru_upload_visualization}
\title{Labguru upload visualization}
\usage{
labguru_upload_visualization(file, title, dataset_id = NULL, name = NULL,
  description = NULL, attach_to_uuid = NULL,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{file}{Single character that is a path to a file}

\item{title}{Single character that is the title for the uploaded file on Labguru}

\item{dataset_id}{Numeric to link the visualization to dataset(s) on Labguru}

\item{name}{Single character that is the name for visualisation link with database}

\item{description}{Single character to describe visualization on Labguru}

\item{server}{}

\item{token}{}
}
\description{
Labguru upload visualization
}
