% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{labguru_valid_token}
\alias{labguru_valid_token}
\title{Labguru Valid Token}
\usage{
labguru_valid_token(token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{token}{Labguru API access token}
}
\value{
logical
}
\description{
Labguru Valid Token
}
\examples{
\dontrun{
labguru_valid_token()
}
}
