% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{labguru_authenticate}
\alias{labguru_authenticate}
\title{Labguru Authenticate}
\usage{
labguru_authenticate(email, password, server = "https://jonathan.labguru.com",
  set_sys = TRUE)
}
\arguments{
\item{email}{Single character email address}

\item{password}{Single character password}

\item{server}{Character - Server URL, "https://jonathan.labguru.com" by default}

\item{set_sys}{Set server and token as system variables LABGURU_SERVER and LABGURU_TOKEN}
}
\value{
acces token (string)
}
\description{
This function authenticates a user by email and password and returns a token and sets the system 
variable LABGURU_TOKEN to that token by default.
}
\examples{
\dontrun{
labguru_authenticate(email    = "my@email.com",
                     password = "mypassword",
                     server   = "https://jonathan.labguru.com")
}
}
