% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{labguru_download_dataset}
\alias{labguru_download_dataset}
\title{Labguru download dataset}
\usage{
labguru_download_dataset(dataset_id, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{dataset_id}{numeric(1) id indicating a dataset on labguru server}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
data frame of labguru dataset
}
\description{
Takes a dataset id and donwloads the dataset.
}
\examples{
\dontrun{
labguru_downlaod_datasets(dataset_id = 1)
}
}
