% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{labguru_ic50_analysis}
\alias{labguru_ic50_analysis}
\title{Labguru ic50 Analysis}
\usage{
labguru_ic50_analysis(plate_id, indir = ".", outdir_plate = "./plate",
  outdir_results = "./results", plates = 2, inhib = NULL,
  normalize = "single", graphics = "mean", img_png = TRUE,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{plate_id}{numeric(1) plate id of the plate from labguru server}

\item{indir}{character(1) argument of ic50::ic50() function with same the name}

\item{outdir_plate}{character(1) directory name where to store plate data}

\item{outdir_results}{character(1) argument of ic50::ic50() function with the name outdir}

\item{plates}{numeric(1) argument of ic50::ic50() function with same the name}

\item{inhib}{numeric(1) argument of ic50::ic50() function with same the name}

\item{normalize}{character(1) argument of ic50::ic50() function with same the name}

\item{graphics}{character(1) argument of ic50::ic50() function with same the name}

\item{img_png}{logical(1) TRUE (default) to create single images from the pdf result, magick package is required for TRUE.}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\description{
Download plate from labguru server and run ic50 analysis.
}
