% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic50.R
\name{labguru_ic50_upload}
\alias{labguru_ic50_upload}
\title{Title}
\usage{
labguru_ic50_upload(table, name, results_dir, img_pdf,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{table}{}

\item{name}{}

\item{results_dir}{}

\item{img_pdf}{}

\item{server}{}

\item{token}{}
}
\description{
Title
}
