% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_link_visualization}
\alias{labguru_link_visualization}
\title{Title}
\usage{
labguru_link_visualization(dataset_id, attachment_id, name,
  description = NULL, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{dataset_id}{}

\item{attachment_id}{}

\item{description}{}

\item{server}{}

\item{token}{}
}
\value{

}
\description{
Title
}
