% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api_functions.R
\name{labguru_list_items}
\alias{labguru_list_items}
\title{Labguru list items}
\usage{
labguru_list_items(url, config = NULL, handle = NULL)
}
\arguments{
\item{url}{}

\item{config}{}

\item{handle}{}

\item{body}{}

\item{encode}{}
}
\value{
parsed object
}
\description{
List
}
