% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_api_functions.R
\name{labguru_post_item}
\alias{labguru_post_item}
\title{Labguru post item}
\usage{
labguru_post_item(url, body, encode = NULL, config = NULL, handle = NULL)
}
\arguments{
\item{url}{url to the labGuru instance API endpoint}

\item{body}{content to be included in POST call}

\item{encode}{encoding of the request (defaults to NULL, should be JSON)}

\item{config}{other configuration options passed on to httr POST}

\item{handle}{handling arguments passed on to httr POST}
}
\value{
parsed object
}
\description{
POST a body to a url with optional arguments. Function includes standard error handling.
}
\examples{
# helper function for POST requests
}
