% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_functions.R
\name{labguru_upload_dataset}
\alias{labguru_upload_dataset}
\title{Labguru Upload Dataset}
\usage{
labguru_upload_dataset(
  dataset,
  name,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN")
)
}
\arguments{
\item{dataset}{data frame you wish to upload to Labguru}

\item{name}{character(1)}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\value{
dataset_id
}
\description{
Upload a dataset to labguru
}
\examples{
\dontrun{
labguru_upload_dataset(dataset    = mtcars,
                       name       = "mtcars")
}
}
