% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_upload_file}
\alias{labguru_upload_file}
\title{Labguru upload file}
\usage{
labguru_upload_file(file, title, description = NULL, attach_to_uuid = NULL,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{file}{Single character that is the path to a file}

\item{title}{Single character as visualization title on Labguru}

\item{description}{Single character to describe visualization on Labguru}

\item{attach_to_uuid}{(optional) single character that is the LG_UUID of the object to which this file should be attached}

\item{server}{Single character indicating the server URL}

\item{token}{Single character access token for API authentication}
}
\value{
list containing uploaded id and url
}
\description{
Upload a file to Labguru
}
\examples{
\dontrun{
data("compensation")

library(ggplot2)
plot <- ggplot(compensation,aes(x = Root, y = Fruit, colour = Grazing)) +
  geom_point(size = 5) +
  xlab("Root Biomass") +
  ylab("Fruit Production") +
  theme_bw()
  
ggsave("grazing.png", plot)

labguru_upload_image(file_path   = "grazing.png",
                     name        = "Grazing",
                     description = "Fruit production versus Root biomass for grazed and ungrazed")
}
}
