% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_upload_visualization}
\alias{labguru_upload_visualization}
\title{Labguru upload visualization}
\usage{
labguru_upload_visualization(file, title, dataset_id = NULL, name = NULL,
  description = NULL, server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{file}{character(1) that is a path to a file}

\item{title}{character(1) that is the title for the uploaded file on Labguru}

\item{dataset_id}{Numeric(1) to link the visualization to dataset(s) on Labguru}

\item{name}{character(1) that is the name for visualisation link with database}

\item{description}{character(1) to describe visualization on Labguru}

\item{server}{character(1) indicating the server URL}

\item{token}{character(1) access token for API authentication}
}
\description{
Upload a visualization to the Labguru server.
}
