% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_file.R
\name{labguru_upload_visualization}
\alias{labguru_upload_visualization}
\title{Labguru upload visualization}
\usage{
labguru_upload_visualization(file_path, name,
  server = Sys.getenv("LABGURU_SERVER"),
  token = Sys.getenv("LABGURU_TOKEN"))
}
\arguments{
\item{file_path}{}

\item{name}{}

\item{server}{}

\item{token}{}
}
\description{
Labguru upload visualization
}
